requirements.txt
torch
torchvision
torchaudio
numpy
pillow
matplotlib
tqdm
scikit-learn
joblib